提交删除会员修改

+++ classes/UserCleanHandlers/UserAddressClear.php
@@ -74,6 +74,11 @@ HTML;
      */
     public function handleClean()
     {
+        $count = $this->handleCount();
+        if (empty($count)) {
+            return true;
+        }
+
         $result = RC_DB::table('user_address')->where('user_id', $this->user_id)->delete();
 
         if ($result) {
