+++ classes/UserCleanHandlers/UserInviteeRecordClear.php
@@ -68,6 +68,11 @@ HTML;
      */
     public function handleClean()
     {
+        $count = $this->handleCount();
+        if (empty($count)) {
+            return true;
+        }
+        
         $result = RC_DB::table('invitee_record')->where('invite_id', $this->user_id)->delete();
 
         if ($result) {
